+++ src/Public/Remove-BurpSuiteScanConfiguration.ps1
@@ -33,3 +33,4 @@ function Remove-BurpSuiteScanConfiguration {
     end {
     }
 }
+