+++ src/Public/Remove-BurpSuiteSiteLoginCredential.ps1
@@ -1,5 +1,6 @@
-function Remove-BurpSuiteSiteApplicationLogin {
+function Remove-BurpSuiteSiteLoginCredential {
     [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'High')]
+    [Alias('Remove-BurpSuiteSiteApplicationLogin')]
     Param (
         [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
         [ValidateNotNullOrEmpty()]
@@ -11,14 +12,14 @@ function Remove-BurpSuiteSiteApplicationLogin {
 
     process {
 
-        $query = _buildMutation -queryName 'DeleteSiteApplicationLogin' -inputType 'DeleteSiteApplicationLoginInput!' -name 'delete_site_application_login' -returnType 'Id' -returnTypeField
+        $query = _buildMutation -queryName 'DeleteSiteLoginCredential' -inputType 'DeleteSiteLoginCredentialInput!' -name 'delete_site_login_credential' -returnType 'Id' -returnTypeField
 
         if ($PSCmdlet.ShouldProcess("BurpSuite", $query)) {
             try {
                 $variables = @{ input = @{} }
                 $variables.input.id = $Id
 
-                $request = [Request]::new($query, 'DeleteSiteApplicationLogin', $variables)
+                $request = [Request]::new($query, 'DeleteSiteLoginCredential', $variables)
 
                 $null = _callAPI -Request $request
             } catch {