+++ unit/tests/public/Update-BurpSuiteSiteLoginCredential.tests.ps1
@@ -1,12 +1,12 @@
 InModuleScope $env:BHProjectName {
-    Describe "Update-BurpSuiteSiteApplicationLogin" {
+    Describe "Update-BurpSuiteSiteLoginCredential" {
         It "should udate site application login" {
             # arrange
             Mock -CommandName _callAPI -MockWith {
                 [PSCustomObject]@{
                     data = [PSCustomObject]@{
-                        update_site_application_login = [PSCustomObject]@{
-                            application_login = [PSCustomObject]@{
+                        update_site_login_credential = [PSCustomObject]@{
+                            login_credential = [PSCustomObject]@{
                                 id = 1
                             }
                         }
@@ -17,12 +17,12 @@ InModuleScope $env:BHProjectName {
             $credentials = New-Object System.Management.Automation.PSCredential ("administrator", $(ConvertTo-SecureString "changeme" -AsPlainText -Force))
 
             # act
-            Update-BurpSuiteSiteApplicationLogin -Id 42 -Label "admin" -Credential $credentials
+            Update-BurpSuiteSiteLoginCredential -Id 42 -Label "admin" -Credential $credentials
 
             # assert
             Should -Invoke _callAPI -ParameterFilter {
-                $Request.OperationName -eq "UpdateSiteApplicationLogin" `
-                    -and $Request.Query -eq 'mutation UpdateSiteApplicationLogin($input:''UpdateSiteApplicationLoginInput!'') { update_site_application_login(input:''$input'') { application_login { id label username } } }' `
+                $Request.OperationName -eq "UpdateSiteLoginCredential" `
+                    -and $Request.Query -eq 'mutation UpdateSiteLoginCredential($input:''UpdateSiteLoginCredentialInput!'') { update_site_login_credential(input:''$input'') { login_credential { id label username } } }' `
                     -and $Request.Variables.Input.id -eq 42 `
                     -and $Request.Variables.Input.label -eq "admin" `
                     -and $Request.Variables.Input.username -eq "administrator" `
@@ -36,8 +36,8 @@ InModuleScope $env:BHProjectName {
                 Mock -CommandName _callAPI -MockWith {
                     [PSCustomObject]@{
                         data = [PSCustomObject]@{
-                            update_site_application_login = [PSCustomObject]@{
-                                application_login = [PSCustomObject]@{
+                            update_site_login_credential = [PSCustomObject]@{
+                                login_credential = [PSCustomObject]@{
                                     id = 1
                                 }
                             }
@@ -46,12 +46,12 @@ InModuleScope $env:BHProjectName {
                 }
 
                 # act
-                Update-BurpSuiteSiteApplicationLogin -Id 42 -Label "admin"
+                Update-BurpSuiteSiteLoginCredential -Id 42 -Label "admin"
 
                 # assert
                 Should -Invoke _callAPI -ParameterFilter {
-                    $Request.OperationName -eq "UpdateSiteApplicationLogin" `
-                        -and $Request.Query -eq 'mutation UpdateSiteApplicationLogin($input:''UpdateSiteApplicationLoginInput!'') { update_site_application_login(input:''$input'') { application_login { id label username } } }' `
+                    $Request.OperationName -eq "UpdateSiteLoginCredential" `
+                        -and $Request.Query -eq 'mutation UpdateSiteLoginCredential($input:''UpdateSiteLoginCredentialInput!'') { update_site_login_credential(input:''$input'') { login_credential { id label username } } }' `
                         -and $Request.Variables.Input.id -eq 42 `
                         -and $Request.Variables.Input.label -eq "admin" `
                         -and $Request.Variables.Input.ContainsKey('username') -eq $false `
@@ -66,8 +66,8 @@ InModuleScope $env:BHProjectName {
                 Mock -CommandName _callAPI -MockWith {
                     [PSCustomObject]@{
                         data = [PSCustomObject]@{
-                            update_site_application_login = [PSCustomObject]@{
-                                application_login = [PSCustomObject]@{
+                            update_site_login_credential = [PSCustomObject]@{
+                                login_credential = [PSCustomObject]@{
                                     id = 1
                                 }
                             }
@@ -78,12 +78,12 @@ InModuleScope $env:BHProjectName {
                 $credentials = New-Object System.Management.Automation.PSCredential ("administrator", $(ConvertTo-SecureString "changeme" -AsPlainText -Force))
 
                 # act
-                Update-BurpSuiteSiteApplicationLogin -Id 42 -Credential $credentials
+                Update-BurpSuiteSiteLoginCredential -Id 42 -Credential $credentials
 
                 # assert
                 Should -Invoke _callAPI -ParameterFilter {
-                    $Request.OperationName -eq "UpdateSiteApplicationLogin" `
-                        -and $Request.Query -eq 'mutation UpdateSiteApplicationLogin($input:''UpdateSiteApplicationLoginInput!'') { update_site_application_login(input:''$input'') { application_login { id label username } } }' `
+                    $Request.OperationName -eq "UpdateSiteLoginCredential" `
+                        -and $Request.Query -eq 'mutation UpdateSiteLoginCredential($input:''UpdateSiteLoginCredentialInput!'') { update_site_login_credential(input:''$input'') { login_credential { id label username } } }' `
                         -and $Request.Variables.Input.id -eq 42 `
                         -and $Request.Variables.Input.ContainsKey('label') -eq $false `
                         -and $Request.Variables.Input.username -eq "administrator" `