+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] - 2021-03-09
+
+### Added
+
+- Implemented new BurpSuite site login credentials type.
+- Added cmdlets for importing site recorded logins.
+- Minor updates and improvements.
+
 ## [1.2.1] - 2020-07-30
 
 ### Added